+++ projects/core/components/textfield/index.ts
@@ -1,5 +1,4 @@
 export * from './textfield';
-export * from './textfield.base';
 export * from './textfield.component';
 export * from './textfield.directive';
 export * from './textfield.options';
